chore: split object handler and mesh handler
